-jacoco plugin added -spring-boot-starter-test dependency added -.travis.yml changed -TestController successful test written -Application.java moved from starter because of this https://docs.spring.io/spring-boot/docs/1.4.0.M3/reference/htmlsingle/#boot-features-testing-spring-boot-applications-detecting-config
