+++ .travis.yml
@@ -1,13 +1,13 @@
 dist: trusty
 
- language: java
+language: java
 
- jdk:
-   - oraclejdk8
+jdk:
+  - oraclejdk8
 
- cache:
-   directories:
-     - $HOME/.m2
+cache:
+  directories:
+   - $HOME/.m2
 
- script:
-    - mvn clean test
+script:
+   - mvn clean test jacoco:report coveralls:report
